requirements.txt
Flask==2.2.2
pdfplumber==0.7.6
openai==0.28.0
python-dotenv==1.0.0